village_py/pyproject.toml
[tool.poetry]
name = "village"
version = "0.1.0"
description = "A basic server backing the village project."
authors = ["Aleksandr Pasechnik <al@megamicron.net>"]
license = "The Unlicense"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.2"
pydantic = "^2.6.4"
pyyaml = "^6.0.1"
markdown = "^3.6"
bleach = "^6.1.0"
pillow = "^10.3.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.9.0"
black = "^24.3.0"
types-pyyaml = "^6.0.12.20240311"
types-markdown = "^3.6.0.20240316"
types-bleach = "^6.1.0.20240331"
isort = "^5.13.2"
ipython = "^8.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
create-user = "village.scripts.create_user:main"
force-reset-password = "village.scripts.force_reset_password:main"
update-thumbnail = "village.scripts.update_thumbnail:main"
